// src/app/shared/model/cust-completion/app-cust-completion-obj.model.ts
export class AppCustCompletionObj{
    AppCustId: number;
    AppId: number;
    CustNo: string;
    ApplicantNo: string;
    CustName: string;
    MrCustTypeCode: string;
    MrCustTypeDescr: string;
    IsFamily: boolean;
    IsShareholder: boolean;
    IsCompletion: boolean;
    MrCustRelationshipCode: string;
    MrCustRelationshipDescr: string;
    SubjectDescr: string;
    BirthPlace: string;
    BirthDt: Date;
    IsOwner: boolean;
    SharePrcnt: number;
    IsSigner: boolean;
    IsActive: boolean;
    EstablishmentDt: Date;
}